+++ routes/web.php
@@ -224,11 +224,11 @@ Route::prefix('pantallainicio/calendario')->group( function(){
   
 //Rutas de Odontologo//
               
-        Route::get('odontologo/nuevo','OdontologoController@nuevoodontologo');
-        Route::post('odontologo/nuevo','OdontologoController@GuardarNuevo');
-        Route::get('/{id}/editar/odontologo','OdontologoController@editarodontologo') ->name('odontologo.editar') -> where('id' ,'[0-9]+');
-        Route::put('/{id}/editar/odontologo','odontologoController@updateodontologo')->name('odontologo.update') -> where('id' ,'[0-9]+');
-        Route::delete('{id}/borrar/odontologo','OdontologoController@destroy') ->name('odontologo.borrar')->where('id','[0-9]+');
+Route::get('odontologo/nuevo','OdontologoController@nuevoodontologo')->middleware('role:admin,secretaria');
+Route::post('odontologo/nuevo','OdontologoController@GuardarNuevo')->middleware('role:admin,secretaria');
+Route::get('/{id}/editar/odontologo','OdontologoController@editarodontologo') ->name('odontologo.editar') -> where('id' ,'[0-9]+')->middleware('role:admin,secretaria');
+Route::put('/{id}/editar/odontologo','odontologoController@updateodontologo')->name('odontologo.update') -> where('id' ,'[0-9]+')->middleware('role:admin,secretaria');
+Route::delete('{id}/borrar/odontologo','OdontologoController@destroy') ->name('odontologo.borrar')->where('id','[0-9]+')->middleware('role:admin,secretaria');
